+++ apps/docs/src/Dialog/Dialog.stories.mdx
@@ -7,7 +7,7 @@ import AlertDialogExample from './AlertDialogExample'
       tags={['autodocs']} 
 />
 
-<Story name="BasicDialog" args={{ header: 'test' }}/>
+<Story name="BasicDialog" args={{}}/>
 
 <Title />
